refactor(consts): floorDiv を consts.FloorDiv へ総称統一しレビュー指摘に対応
2度のレビューで重複を指摘された floorDivInt(query) と floorDiv(overworld) を、
両者が import する leaf の consts へ総称 FloorDiv[T ~int] として単一出典化する。
BandOriginY の単一出典化と同じ形で、依存方向を理由にした重複コメントも消える。
テストは関数の住所へ移し Chunk と int の両インスタンスを固定する。

あわせて軽微な指摘に対応する。
- overworld_map.go: overworldMapCell の doc が古い行/高さ記述だったのを実態の
  列/幅へ直し、マジックナンバー +3 を overworldMapMarginCols へ命名
- band_test.go: 帯外へリベースされた探索済みキーの破棄を NotContains で明示
