+++ internal/overworld/road_geometry.go
@@ -46,7 +46,7 @@ func (s roadSeg) tileSpan(chunkW, chunkH consts.Tile) (fixed, lo, hi consts.Tile
 // 結ぶ2本だけ。road.go の舗装と scatter.go の回避が同じ結線を共有する。ホットパスで呼ぶので、
 // スライスを確保せず固定長配列で返す。集落は Y 方向のリージョンに並ぶので region は c.Y から引く。
 func crossingRoads(runSeed uint64, c consts.Coord[consts.Chunk], cols consts.Chunk) [2][2]consts.Coord[consts.Chunk] {
-	r := floorDiv(c.Y, settlementPlacement.Spacing)
+	r := consts.FloorDiv(c.Y, settlementPlacement.Spacing)
 	var pairs [2][2]consts.Coord[consts.Chunk]
 	for i := range 2 {
 		pr := r - 1 + consts.Chunk(i)
